Move environment variable exports for file locations to iso/exports.
Maintenance of create_box to add new environment variables for ISO locations
would result in constant change unrelated to actual box creation. Instead, pull
the exports where they can be sourced instead for the needs of the local
boxcutter. The downside is that ISO existence verification can not happen as
a pre-check before starting a box build, requiring initial vigilance of stdout
output.